fix: add missing ios bridge file

+++ android/src/main/java/com/ravij4057/plugins/shake/MyShakePlugin.java
@@ -40,7 +40,7 @@ public class MyShakePlugin extends Plugin implements SensorEventListener {
         float y = event.values[1];
         float z = event.values[2];
         mAccelLast = mAccelCurrent;
-        mAccelCurrent = (float) Math.sqrt(x * x + y * y + z * z);
+        mAccelCurrent = (float) Math.sqrt((double) (x * x + y * y + z * z));
         float delta = mAccelCurrent - mAccelLast;
         mAccel = mAccel * 0.9f + delta;
         if (mAccel > 12) {
